Synch With Bill's C5535 DSP Shield Changes
Includes HAL changes for DSP Shield as well as hardware\tools\c5000\bin
copy of avrdude

+++ hardware/C5000/libraries/SPI/SPI.h
@@ -110,8 +110,8 @@ class SPI_Class
         void setLoopBackMode(int value);
         void setBitOrder(int order);
         int transfer (int value);
-        int write (unsigned long buffer[], int length);
-        int read (unsigned long buffer[], int length);
+        int write (unsigned int buffer[], int length);
+        int read (unsigned int buffer[], int length);
 } ;
 
 extern SPI_Class SPI;
